requirements.txt
tweepy
newsapi-python
schedule